add support for images and other documents
If a web-server is running on the same host as this bridge, files
uploaded from within telegram can now be shared by sending links
to these files to irc.

To enable this feature, set baseURL and localPath in telegram.cfg.

Make sure that the directory listing for the specified URL is
disabled (probably achieved by chmodding it to 711 or touching
an empty index.html).